pyproject.toml
[tool.poetry]
name = "query-pager"
version = "0.1.0"
description = "Cursor-based pagination and CEL filtering for Django and SQLAlchemy"
authors = ["Wonguk Jung <eric.jung@blitzsoft.co.kr>"]
license = "MIT"
readme = "README.md"
packages = [{include = "query_pager", from = "src"}]
homepage = "https://github.com/blitzsoft-kr/query-pager"
repository = "https://github.com/blitzsoft-kr/query-pager"
keywords = ["pagination", "filtering", "django", "sqlalchemy", "cel", "cursor"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^2.0"
cel-python = ">=0.3.0,<0.4"
typing-extensions = "^4.0"
# Align with cel-python's parser requirement to avoid resolver conflicts
lark = ">=0.12,<0.13"

# Optional dependencies
sqlalchemy = {version = ">=1.3.11", optional = true}
django = {version = ">=4.0", optional = true}

[tool.poetry.extras]
sqlalchemy = ["sqlalchemy"]
django = ["django"]
all = ["sqlalchemy", "django"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21"
pytest-django = "^4.5"
pytest-cov = "^4.0"
black = "^23.0"
ruff = "^0.1"
mypy = "^1.0"
aiosqlite = "^0.19.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
DJANGO_SETTINGS_MODULE = "django_settings"
pythonpath = ["tests"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "celpy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlakeyset.*"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
